+++ src/deadline/client/job_bundle/_utils.py
@@ -5,9 +5,6 @@ Provides deadline_yaml_dump, which works like pyyaml's safe_dump,
 but saves multi-line strings with the "|" style.
 """
 
-import json
-import os
-from typing import Any
 import yaml
 from yaml.emitter import Emitter
 from yaml.representer import SafeRepresenter
@@ -81,24 +78,3 @@ def deadline_yaml_dump(data, stream=None, **kwds):
     strings with the "|" style and defaults to sort_keys=False.
     """
     return yaml.dump_all([data], stream, Dumper=DeadlineDumper, sort_keys=False, **kwds)
-
-
-def save_yaml_or_json_to_file(
-    bundle_dir: str,
-    filename: str,
-    file_type: str,
-    data: Any,
-) -> None:
-    """
-    Saves data as either a JSON or YAML file depending on the file_type provided. Useful for saving
-    job bundle data files which can be in either format.
-    """
-    with open(
-        os.path.join(bundle_dir, f"{filename}.{file_type.lower()}"), "w", encoding="utf8"
-    ) as f:
-        if file_type == "YAML":
-            deadline_yaml_dump(data, f)
-        elif file_type == "JSON":
-            json.dump(data, f, indent=2)
-        else:
-            raise RuntimeError(f"Unexpected file type '{file_type}' in job bundle:\n{bundle_dir}")
