+++ src/deadline/client/ui/job_bundle_submitter.py
@@ -19,7 +19,7 @@ from ..job_bundle.loader import (
     read_yaml_or_json_object,
     validate_directory_symlink_containment,
 )
-from ..job_bundle._utils import save_yaml_or_json_to_file
+from ..job_bundle.saver import save_yaml_or_json_to_file
 from ..job_bundle.parameters import (
     JobParameter,
     apply_job_parameters,
